requirements.txt
requests>=2.28.1
dnspython>=2.2.1
tqdm>=4.64.1
gensim>=4.2.0
pyyaml>=6.0
torch>=1.13.0
